requirements.txt
discord.py
python-dotenv
yt-dlp
youtube-search-python
lyricsgenius